backend/requirements.txt
parameters-validation
flask
python-jose[cryptography]

